refactor: remove unused code and streamline memory access in FusePhaseProcessor and related classes

+++ machine/src/main/java/com/fpetrola/oozx/Memory.java
@@ -253,7 +253,7 @@ public class Memory extends DefaultRAMHolder implements ZxModule {
   // Read a byte from memory
   public void readByte(int address, Ula ula) {
     if (mapRead[address >>> PAGE_SIZE_LOGARITHM].contended) {
-      zxClock.addTStates(ula.contention[(int) zxClock.getTStates()], "ula readbyte");
+      zxClock.addTStates(ula.contention[zxClock.getTStates()], "ula readbyte");
     }
 //      tStatesHolder.tstates += 3;
 //    mapping.getPage().get(address & PAGE_SIZE_MASK);
@@ -267,7 +267,7 @@ public class Memory extends DefaultRAMHolder implements ZxModule {
   // Write a byte to memory
   public void writeByte(int address, byte b, Ula ula) {
     if (mapWrite[address >>> PAGE_SIZE_LOGARITHM].contended) {
-      zxClock.addTStates(ula.contention[(int) zxClock.getTStates()], "ula writebyte");
+      zxClock.addTStates(ula.contention[zxClock.getTStates()], "ula writebyte");
     }
 //      tStatesHolder.tstates += 3;
 //    writeByteInternal(address, b);
